Search Post

+++ src/posts/views.py
@@ -1,17 +1,19 @@
 from urllib.parse import urlencode, quote_plus
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.db.models import Q
+from django.utils import timezone
 from django.contrib import messages
 from .forms import PostForm
 from .models import Post
 
 def post_create(request):
-	# if not request.user.is_staff or not request.user.is_superuser:
-	# 	raise Http404
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 
-	# if not request.user.is_auhenticated:
-	# 	raise Http404
+	if not request.user.is_auhenticated:
+		raise Http404
 
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
@@ -27,8 +29,8 @@ def post_create(request):
 	return render(request, "post_form.html", context)
 
 def post_update(request, slug=None):
-	# if not request.user.is_staff or not request.user.is_superuser:
-	# 	raise Http404
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
@@ -47,6 +49,9 @@ def post_update(request, slug=None):
 
 def post_detail(request, slug):
 	instance = get_object_or_404(Post, slug=slug)
+	if instance.publish > timezone.now().date() or instance.draft:
+		if not request.user.is_staff or not request.user.is_superuser:
+			raise Http404
 	share_string = quote_plus(instance.content)
 	context = {
 		"title": instance.title,
@@ -56,7 +61,19 @@ def post_detail(request, slug):
 	return render(request, "post_detail.html", context)
 
 def post_list(request):
-	queryset_list = Post.objects.all() #.order_by("-timestamp")
+	today = timezone.now().date()
+	queryset_list = Post.objects.active()  #filter(publish__lte=timezone.now())#active() #filter(draft=False)#.order_by("-timestamp")
+	if request.user.is_staff or not request.user.is_superuser:
+		queryset_list = Post.objects.all()
+
+	query = request.GET.get("q")
+	if query:
+		queryset_list = queryset_list.filter(
+			Q(title__icontains=query) |
+			Q(content__icontains=query) |
+			Q(user__first_name__icontains=query) |
+			Q(user__last_name__icontains=query) 
+			).distinct()
 	paginator = Paginator(queryset_list, 3)
 	page_request_var = "page"
 	page = request.GET.get(page_request_var)
@@ -75,8 +92,8 @@ def post_list(request):
 	return render(request, "post_list.html", context)
 
 def post_delete(request, slug=None):
-	# if not request.user.is_staff or not request.user.is_superuser:
-	# 	raise Http404
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully Deleted")
